+++ src/scene/primitive.h
@@ -11,8 +11,8 @@ struct CudaPrimitive {
 
   DEVICE bool intersect(Ray& r, CudaIntersection* i, Vector3D* vertices, Vector3D* normals);
 
-  size_t i_p1, i_p2, i_p3;
-  size_t i_n1, i_n2, i_n3;
+  uint32_t i_p1, i_p2, i_p3;
+  uint32_t i_n1, i_n2, i_n3;
   
   uint32_t bsdf_idx;
 };
